Implementing hydra
First step into implementing the website with hydra (video synth) with the intent of porting the original p5.js sketch to the instance mode of p5.js which can run within hydra, on the same canvas. This commit only implements a hydra canvas behind the textbox and button, but it covers the fire from the original p5.js sketch. The forgiving message and alpha of bubbles have also been slightly changed.